Update filter file parsing to strip newline characters
The readlines() string method leaves in newline characters. This
behavior ends up preventing matches between the newline-delimited
filter strings read from a file and the bin object path strings. By
reading in the full filter file with read(), and then splitting
on the newline characters with splitlines() (both \n and \r\n are
supported), the trailing newline characters are eliminated and
filepath patterns are matched as expected.